requirements.txt
PyQt5~=5.15.10
QtPy~=2.4.1
gitpython~=3.1.18
